update
Former-commit-id: 26736f2e3363635f16c4a81f454f2ed6fedc82f6

+++ user/yliu/deviation.c
@@ -38,7 +38,7 @@ float sdeviation(float *temp, int nfw)
 	data[i]=(data[i]-m)*(data[i]-m);
     }
     sd=mean(data,nfw);
-    sd=sqrt(sd);
+    sd=sqrtf(sd);
     return sd;
 }
